Switched to patch sets and added calling context
I switched to using patch sets, since I think it is easier to understand and
more concise to specify (don't need a 'triples =').

I also provided a context which enforces ordering across all functions that are
part of the context.

I have not yet updated the documentation; I'll update it if this patch is
accepted.